+++ index.js
@@ -1,3 +1,1 @@
-const log = require('@dashboard/logger')
-
-log.info('Hello World', module)
+require('./@dashboard/app')